rename infographic/routes.js to infographic_link.js for clarity
